Fix: Library dependency issue in pdfClient_test which is causing MTS failure on android T and below devices.
Root Cause - The actual pdfclient library compiles pdfium statically
while the test was using a shared library which goes to a build from
platform and that build had older version on lower android versions
which has some methods missing which are required by newly added tests.
Thus causing error in MTS tests on android T and below.

TODO : Need to add libpdfclient as dependency in test directly, which
will be added in subsequent cl. Currently header files in the directory
is not detecting which need to be fixed for this.

Bug: 390327689
Flag: EXEMPT Test Only
Test: Ran MTS manually on android R,S,T,U and V
Change-Id: I5a2ea6c4323228daffbadbb4fdaf808049953bcb