Owned photos unit test fix
The DisableCompatChanges check was added to the test
testFetchMediaGrantsCount. It should be added only of B+ versions but
test is run for S+, so removing the check. The flag is already disabled,
that should be enough to disable owned photos check

Bug: 390349036
Test: com.android.providers.media.photopicker.v2.PickerDataLayerV2Test#testFetchMediaGrantsCount
Flag: com.android.providers.media.flags.revoke_access_owned_photos
Change-Id: Ic8efac957454e391ae45f08fa642be76c945fd35