Fix NPE due to race condition in SearchDataService
Property initialisers and init blocks are executed
in the order they appear in the code. So, in SearchDataServiceImpl,
there was a race condition where the collection block of
dataservice.availableproviders executed before _userSearchStateInfo was
initialized, causing a NPE.

This change fixes the ordering of the init block and
_userSearchStateInfo definition.

Bug: 391605301
Test: The repro rate is extremely low so this is difficult to test. Will
keep an eye on the crash clusters.
Flag: com.android.providers.media.flags.enable_photopicker_search

Change-Id: If5fd351df54fa0315b96f65541f34e4273f8e3d6