Default folders: Add upper bound to exclusion list size
Ensure that the exclusion list is not bigger than the default list. If
this check fails, log an error and set the exclusion list to be empty.
This check is important as OEMs can control the exclusion list and we
want to put an upper bound on the exclusionList.contains() call.

Also put the exclusion list logic behind the shared preference check.

Bug: 381310841
Test: atest com.android.providers.media.MediaProviderTest
Flag: com.android.providers.media.flags.enable_exclusion_list_for_default_folders
Change-Id: Ieb90fc09c5fbf46f5fb13e4f4c978819d8a316dd