Change catgory grid route to albumgrid
With this change the route for CategoryGridFeature is updated to be same
as AlbumGridFeature since both the features will never be enabled
together.
With this when category feature is enabled it is made sure photopicker
is started on collections tab whenever Album is selected as start
destination.

Recording:
https://drive.google.com/file/d/1I6k_Lx2jFY8NEGKJFdE2AtztlQs0AqOv/view?usp=sharing

Bug: 392596492
Test: atest CategoryGridFeatureTest
Flag: EXEMPT bug fix
Change-Id: I33934e5e517afbf700a942825ad44c22ce0b731f