Fix for legacy storage app op
There is a bug in AppOpsManager that keeps legacy storage granted even
when an app updates its targetSdkVersion from value <30 to >=30.
If an app upgrades from targetSdk 29 to targetSdk 33, legacy storage
remains granted and in targetSdk 33, app are required to replace R_E_S
with R_M_*. If an app updates its manifest with R_M_*, permission check
in MediaProvider will look for R_E_S and will not grant read access as
the app would be still treated as legacy.

Test: atest PermissionUtilsTest
Bug: 315914683
Flag: EXEMPT bug fix
Change-Id: I22d0a329fbea5afe9dea2bc5179596db7a2ef402