Clear media sets cache and media sets content cache for a given categoryId and its authority
The current implementation erases the complete cache for both media sets
and media set content. It is fine for the MVP since we only have one
category at the moment but in case we add more in the future, this
should not be the behaviour.

Test: verified manually, fixed corresponding tests

Bug: 394050877

Flag: com.android.providers.media.flags.enable_photopicker_search
Change-Id: Idd8980c0e573376df8100b44102dcc01b5e8e992