Annotate lambdas mocking UserManager.getUserProperties
Kotlin 2 compiler synthetically converts lambdas to private methods
in the test class. When Junit runs, it tries to enumerate all of
the methods including private methods. This would fail if
the lambda was capturing a class (e.g UserProperties)
that doesn't exist on the build the test is running on.

This change is using @JvmSerializableLambda annotation in these
instances to emit synthetic class vs synthetic private method.

Bug: 391761542
Change-Id: If5ea889ce4874e4a3662438843717cad59722440
Test: atest PhotoPickerTests
Flag: EXEMPT bug fix