Defer opening picker DB to the first use
Currently, picker DB is being opened in the initialization path of
MediaProvider which is discouraged operation. DB open could take time to
happen, especially if DB upgrade is to occure, which could significantly
increase MediaProvider initialization.

Instead, we deferred opening the DB to the first use of the DB.
getWritableDatabase is making sure that the DB is opened once, by
caching the DB opened on the first call and returning the same DB on any
subsequent get.

Test: atest PickerDbFacadeTest
Bug: 399347278
Flag: EXEMPT bug fix
Change-Id: Ib57a0f1716349ad5a8c71977035835b69247f2ce