Guard against race condition in CheckFdAccess
We have been seeing reports where race condition was happening as a
result of node deletion, while trying to check if an app has access to
it.

With this fix, we are making sure that the node won't be deleted after
it was looked up and before checking the needed info on it.

Bug: 374685576
Bug: 339567926
Bug: 396594195
Test: atest CtsMediaProviderTranscodeTests
Flag: EXEMPT bugfix
Change-Id: I91caba1ee65f7b6a9827253e9256f88bf16fb854